temporal_res_model

+++ src/losses.py
@@ -14,7 +14,7 @@ def masked_mse_loss(
     y_true: (B, 96, T)
     mask: (B, 96) where 1 = masked channel
     """
-    mask_exp = mask.unsqueeze(-1)  # (B,96,1)
+    mask_exp = mask.unsqueeze(-1).to(dtype=y_hat.dtype)  # (B,96,1)
 
     diff_sq = (y_hat - y_true).pow(2)
     masked_diff_sq = diff_sq * mask_exp
@@ -32,7 +32,7 @@ def nmse_masked(
     """
     NMSE = sum((y_hat-y)^2 * mask) / sum((y^2) * mask), over all masked channels and time.
     """
-    mask_exp = mask.unsqueeze(-1)  # (B,96,1)
+    mask_exp = mask.unsqueeze(-1).to(dtype=y_hat.dtype)  # (B,96,1)
 
     num = ((y_hat - y_true).pow(2) * mask_exp).sum()
     den = (y_true.pow(2) * mask_exp).sum().clamp(min=eps)
